* Big giant load of updates and adjustments

+++ public_html/includes/autoloader.inc.php
@@ -30,33 +30,21 @@
 
 			// Submodules
 			case (preg_match('#^job_#', $class)):
+				$file = match(strtok($class, '_')) {
+					'job' => FS_DIR_APP . 'includes/modules/jobs/' . $class . '.inc.php',
+					'tm' => FS_DIR_APP . 'includes/modules/translation/' . $class . '.inc.php',
+				};
 
 				// Patch modules for PHP 8.2 Compatibility
-				if (version_compare(PHP_VERSION, 8.2, '>=')) {
-
-					$search_replace = [
-						'#^job_.*$#' => 'app://includes/modules/jobs/$1.inc.php',
-					];
-
-					$file = preg_replace(array_keys($search_replace), array_values($search_replace), $class);
-
-					if (is_file($file)) {
-						$source = file_get_contents($file);
-
-						if (!preg_match('#class [a-zA-Z0-9_-]+ extends abs_module#', $source)) {
-							$source = preg_replace('#(class [a-zA-Z0-9_-]+) *\{#', '$1 extends abs_module {', $source);
-							file_put_contents($file, $source);
-						}
+				if (version_compare(PHP_VERSION, 8.2, '>=') && is_file($file)) {
+					$source = file_get_contents($file);
+					if (!preg_match('#\#\[AllowDynamicProperties\]#', $source)) {
+						$source = preg_replace('#([ \t]*)class [a-zA-Z0-9_-]+\s*?\{(\n|\r\n?)#', '$1#[AllowDynamicProperties]$2$0', $source);
+						file_put_contents($file, $source);
 					}
 				}
 
-				switch (true) {
-
-					case (preg_match('#^job_#', $class)):
-						require 'app://includes/modules/jobs/' . $class . '.inc.php';
-						break;
-				}
-
+				require $file;
 				break;
 
 			// References
